Shims cleanup
